Changes made to make login/logout information dynamic, and also added refactoring changes to take routes out of app.py, and add to their dedicated file within the /routes/ directory.
