+++ nvsvocprez/app.py
@@ -1,7 +1,7 @@
 import logging
 import json
 
-from routes import about_page, contact_page, collection_page
+from routes import about_page, contact_page, collection_pages, index_page, scheme_pages
 from authentication import login
 
 
@@ -36,12 +36,16 @@ from utils import (
     get_collection_query, 
     get_ontologies
 )
+
 from pyldapi import Renderer, ContainerRenderer, DisplayProperty
 from config import SYSTEM_URI, DATA_URI, PORT
 from rdflib import Graph, URIRef
 from rdflib import Literal as RdfLiteral, Namespace
 from rdflib.namespace import DC, DCTERMS, ORG, OWL, RDF, RDFS, SKOS, VOID
 
+
+#### Initial Setup ####
+# NOTE: This should be refactored into a function/Class
 api_home_dir = Path(__file__).parent
 with open("api_doc_config.json", "r") as config_file:
     doc_config = json.load(config_file)
@@ -80,970 +84,21 @@ acc_dep_map = {
     None: "",
 }
 
+#### Add the imported routes to the API ####
+# NOTE: To be refactored into a function, or part of a Class.
+api.include_router(index_page.router)
+api.include_router(about_page.router)
+api.include_router(contact_page.router)
+api.include_router(collection_pages.router)
+api.include_router(login.router)
+api.include_router(scheme_pages.router)
+
+
 # This function is here temporarily, when all routes are inside the /routes/ folder
 # then this function will need to go inside /routes/utils/
 def get_user_status(request: Request) -> str:
     """Check the current session to return name of logged in user, if logged in."""
     return request.session['user']['nickname'] if 'user' in request.session else 'Not Logged in'
-
-
-@api.get("/", include_in_schema=False)
-@api.head("/", include_in_schema=False)
-def index(request: Request):
-    dcat_file = api_home_dir / "dcat.ttl"
-    sdo_file = api_home_dir / "sdo.ttl"
-
-    class DatasetRenderer(Renderer):
-        def __init__(self):
-            self.instance_uri = SYSTEM_URI
-            self.label = "NERC Vocabulary Server Content"
-            self.comment = (
-                "The NVS gives access to standardised and hierarchically-organized vocabularies. It is "
-                "managed by the British Oceanographic Data Centre at the National Oceanography Centre "
-                "(NOC) in Liverpool and Southampton, and receives funding from the Natural Environment "
-                "Research Council (NERC) in the United Kingdom. Major technical developments have also "
-                "been funded by European Union's projects notably the Open Service Network for Marine "
-                "Environmental Data (NETMAR) programme, and the SeaDataNet and SeaDataCloud projects."
-            )
-            super().__init__(
-                request,
-                self.instance_uri,
-                {"dcat": dcat, "sdo": sdo},
-                "dcat",
-            )
-
-        def render(self):
-            if self.profile == "dcat":
-                if self.mediatype == "text/html":
-                    return templates.TemplateResponse(
-                        "index.html", {"request": request,
-                        "login_status" : get_user_status(request),
-                        "logged_in_user" : get_user_status(request)}
-                    )
-                else:  # all other formats are RDF
-                    if self.mediatype == "text/turtle":
-                        return Response(
-                            open(dcat_file).read().replace("xxx", self.instance_uri),
-                            headers={"Content-Type": "text/turtle"},
-                        )
-                    else:
-                        g = Graph().parse(
-                            data=open(dcat_file)
-                            .read()
-                            .replace("xxx", self.instance_uri),
-                            format="turtle",
-                        )
-                        return Response(
-                            content=g.serialize(format=self.mediatype),
-                            headers={"Content-Type": self.mediatype},
-                        )
-            elif self.profile == "sdo":
-                if self.mediatype == "text/turtle":
-                    return Response(
-                        open(sdo_file).read().replace("xxx", self.instance_uri),
-                        headers={"Content-Type": "text/turtle"},
-                    )
-                else:
-                    g = Graph().parse(
-                        data=open(sdo_file).read().replace("xxx", self.instance_uri),
-                        format="turtle",
-                    )
-                    return Response(
-                        content=g.serialize(format=self.mediatype),
-                        headers={"Content-Type": self.mediatype},
-                    )
-
-            alt = super().render()
-            if alt is not None:
-                return alt
-
-    return DatasetRenderer().render()
-
-
-
-@api.get("/scheme/", **paths["/scheme/"]["get"])
-@api.head("/scheme/", include_in_schema=False)
-def conceptschemes(request: Request):
-    class ConceptSchemeRenderer(ContainerRenderer):
-        def __init__(self):
-            self.instance_uri = SYSTEM_URI
-            self.label = "NVS Thesauri"
-            self.comment = (
-                "SKOS concept schemes managed by the NERC Vocabulary Server. A concept scheme can be "
-                "viewed as an aggregation of one or more SKOS concepts. Semantic relationships (links) "
-                "between those concepts may also be viewed as part of a concept scheme. A concept scheme "
-                "is therefore useful for containing the concepts registered in multiple concept "
-                "collections that relate to each other as a single semantic unit, such as a thesaurus."
-            )
-            super().__init__(
-                request,
-                self.instance_uri,
-                {"nvs": nvs},
-                "nvs",
-            )
-
-        def _render_sparql_response_rdf(self, sparql_response):
-            if sparql_response[0]:
-                return Response(
-                    '<?xml version="1.0" encoding="UTF-8"?>\n'.encode() + sparql_response[1]
-                    if "xml" in self.mediatype else sparql_response[1],
-                    headers={"Content-Type": self.mediatype}
-                )
-            else:
-                return PlainTextResponse(
-                    "There was an error obtaining the Concept RDF from the Triplestore",
-                    status_code=500,
-                )
-
-        def render(self):
-            if self.profile == "nvs":
-                if self.mediatype == "text/html":
-                    conceptschemes = cache_return(
-                        collections_or_conceptschemes="conceptschemes"
-                    )
-
-                    if request.query_params.get("filter"):
-
-                        def concat_vocab_fields(vocab):
-                            return (
-                                f"{vocab['id']['value']}"
-                                f"{vocab['prefLabel']['value']}"
-                                f"{vocab['description']['value']}"
-                            )
-
-                        conceptschemes = [
-                            x
-                            for x in conceptschemes
-                            if request.query_params.get("filter")
-                            in concat_vocab_fields(x)
-                        ]
-
-                    return templates.TemplateResponse(
-                        "conceptschemes.html",
-                        {
-                            "request": request,
-                            "uri": self.instance_uri,
-                            "label": self.label,
-                            "comment": self.comment,
-                            "conceptschemes": conceptschemes,
-                            "profile_token": "nvs",
-                        },
-                    )
-                elif self.mediatype in RDF_MEDIATYPES:
-                    q = """
-                        PREFIX skos: <http://www.w3.org/2004/02/skos/core#>
-                        PREFIX dc: <http://purl.org/dc/terms/>
-                        PREFIX owl: <http://www.w3.org/2002/07/owl#>
-                        PREFIX xsd: <http://www.w3.org/2001/XMLSchema#>
-                        
-                        CONSTRUCT {
-                            ?cs a skos:ConceptScheme ;
-                                dc:alternative ?alt ;
-                                dc:creator ?creator ;
-                                dc:date ?modified ;
-                                dc:publisher ?publisher ;
-                                dc:title ?title ;
-                                owl:versionInfo ?version ;
-                                skos:hasTopConcept ?tc ;
-                                skos:altLabel ?al ;
-                                dc:description ?description ;
-                                skos:prefLabel ?pl .
-                        }
-                        WHERE {
-                            ?cs a skos:ConceptScheme ;
-                                dc:alternative ?alt ;
-                                dc:creator ?creator ;
-                                dc:date ?m ;
-                                dc:publisher ?publisher ;
-                                dc:title ?title ;
-                                owl:versionInfo ?version ;
-                            .
-                            BIND (STRDT(REPLACE(STRBEFORE(?m, "."), " ", "T"), xsd:dateTime) AS ?modified)
-
-                            OPTIONAL {?cs skos:hasTopConcept ?tc .}
-                            OPTIONAL { ?cs skos:altLabel ?al . }
-                            {
-                                ?cs dc:description ?description .
-                                FILTER(lang(?description) = "en" || lang(?description) = "")
-                            }
-                            {
-                                ?cs skos:prefLabel ?pl .
-                                FILTER(lang(?title) = "en" || lang(?pl) = "")
-                            }
-                        }
-                        """
-                    return self._render_sparql_response_rdf(sparql_construct(q, self.mediatype))
-            elif self.profile == "mem":
-                collections = []
-                for c in cache_return(collections_or_conceptschemes="conceptschemes"):
-                    collections.append(
-                        {
-                            "uri": c["uri"]["value"],
-                            "systemUri": c["systemUri"]["value"],
-                            "label": c["prefLabel"]["value"],
-                        }
-                    )
-
-                if self.mediatype == "text/html":
-                    return templates.TemplateResponse(
-                        "container_mem.html",
-                        {
-                            "request": request,
-                            "uri": self.instance_uri,
-                            "label": self.label,
-                            "collections": collections,
-                            "profile_token": "nvs",
-                        },
-                    )
-                elif self.mediatype == "application/json":
-                    return [
-                        {"uri": c["uri"], "label": c["prefLabel"]} for c in collections
-                    ]
-                else:  # all other available mediatypes are RDF
-                    g = Graph()
-                    container = URIRef(self.instance_uri)
-                    g.add((container, RDF.type, RDF.Bag))
-                    g.add((container, RDFS.label, RdfLiteral(self.label)))
-                    for c in collections:
-                        g.add((container, RDFS.member, URIRef(c["uri"])))
-                        g.add((URIRef(c["uri"]), RDFS.label, RdfLiteral(c["label"])))
-                    return Response(
-                        g.serialize(format=self.mediatype),
-                        media_type=self.mediatype
-                    )
-            elif self.profile == "contanno":
-                if self.mediatype == "text/html":
-                    return templates.TemplateResponse(
-                        "container_contanno.html",
-                        {
-                            "request": request,
-                            "uri": self.instance_uri,
-                            "label": self.label,
-                            "comment": self.comment,
-                            "profile_token": "nvs",
-                        },
-                    )
-                else:  # all other available mediatypes are RDF
-                    g = Graph()
-                    container = URIRef(self.instance_uri)
-                    g.add((container, RDF.type, RDF.Bag))
-                    g.add((container, RDFS.label, RdfLiteral(self.label)))
-                    c = (
-                        "This object is a container that contains a number of members. See other profiles of this "
-                        "object to see those members."
-                    )
-                    c += self.comment
-                    g.add((container, RDFS.comment, RdfLiteral(c)))
-                    return Response(
-                        g.serialize(format=self.mediatype),
-                        media_type=self.mediatype
-                    )
-            alt = super().render()
-            if alt is not None:
-                return alt
-
-    return ConceptSchemeRenderer().render()
-
-
-@api.get("/collection/{collection_id}", include_in_schema=False)
-@api.get("/collection/{collection_id}/", include_in_schema=False)
-@api.head("/collection/{collection_id}" , include_in_schema=False)
-@api.head("/collection/{collection_id}/" , include_in_schema=False)
-def collection_no_current(request: Request, collection_id):
-    return RedirectResponse(url=f"/collection/{collection_id}/current/")
-
-
-@api.get("/collection/{collection_id}/current/", **paths["/collection/{collection_id}/current/"]["get"])
-@api.get("/collection/{collection_id}/current/{acc_dep_or_concept}/",**paths["/collection/{collection_id}/current/{acc_dep_or_concept}/"]["get"])
-@api.head("/collection/{collection_id}/current/" , include_in_schema=False)
-@api.head("/collection/{collection_id}/current/{acc_dep_or_concept}/" , include_in_schema=False)
-def collection(request: Request, collection_id, acc_dep_or_concept: str = None):
-
-    if not exists_triple(request.url.path) and acc_dep_or_concept not in ["accepted", "deprecated", "all"]:
-      raise HTTPException(status_code=404)
-
-    if acc_dep_or_concept not in ["accepted", "deprecated", "all", None]:
-        # this is a call for a Concept
-        return concept(request)
-
-    class CollectionRenderer(Renderer):
-        def __init__(self):
-            self.alt_profiles = get_alt_profiles()
-            self.ontologies = get_ontologies()
-            self.instance_uri = f"{DATA_URI}/collection/{collection_id}/current/"
-
-            profiles = {"nvs": nvs, "skos": skos, "vocpub": vocpub, "dd": dd}
-            for collection in cache_return(collections_or_conceptschemes="collections"):
-                if collection["id"]["value"] == collection_id:
-                    if collection.get("conforms_to"):
-                        profiles.update(
-                            get_alt_profile_objects(
-                                collection = collection,
-                                alt_profiles= self.alt_profiles, 
-                                ontologies = self.ontologies
-                            )
-                        )
-
-            super().__init__(
-                request,
-                self.instance_uri,
-                profiles,
-                "nvs",
-            )
-
-        def _render_sparql_response_rdf(self, sparql_response):
-            if sparql_response[0]:
-                return Response(
-                    '<?xml version="1.0" encoding="UTF-8"?>\n'.encode() + sparql_response[1]
-                    if "xml" in self.mediatype else sparql_response[1],
-                    headers={"Content-Type": self.mediatype}
-                )
-            else:
-                return PlainTextResponse(
-                    "There was an error obtaining the Concept RDF from the Triplestore",
-                    status_code=500,
-                )
-
-        def _get_collection(self):
-            for collection in cache_return(collections_or_conceptschemes="collections"):
-                if collection["id"]["value"] == collection_id:
-                    return collection
-
-        def _get_concepts(self):
-            q = """
-                PREFIX dcterms: <http://purl.org/dc/terms/>
-                PREFIX skos: <http://www.w3.org/2004/02/skos/core#>
-                SELECT DISTINCT ?c ?systemUri ?id ?pl ?def ?date ?dep
-                WHERE {
-                        <xxx> skos:member ?c .
-                        BIND (STRBEFORE(STRAFTER(STR(?c), "/current/"), "/") AS ?id)
-                        BIND (STRAFTER(STR(?c), ".uk") AS ?systemUri)
-
-                        acc_dep
-                        
-                        OPTIONAL {
-                            ?c <http://www.w3.org/2002/07/owl#deprecated> ?dep .
-                        }
-                        ?c skos:prefLabel ?pl ;
-                             skos:definition ?def ;
-                             dcterms:date ?date .
-
-                        FILTER(lang(?pl) = "en" || lang(?pl) = "") 
-                        FILTER(lang(?def) = "en" || lang(?def) = "")                    
-                }
-                ORDER BY ?pl
-                """.replace(
-                "xxx", self.instance_uri
-            ).replace(
-                "acc_dep", acc_dep_map.get(acc_dep_or_concept)
-            )
-
-            sparql_result = sparql_query(q)
-            if sparql_result[0]:
-                return [
-                    {
-                        "uri": concept["c"]["value"],
-                        "id": concept["id"]["value"],
-                        "systemUri": concept["systemUri"]["value"],
-                        "prefLabel": concept["pl"]["value"].replace("_", " "),
-                        "definition": concept["def"]["value"].replace("_", "_ "),
-                        "date": concept["date"]["value"][0:10],
-                        "deprecated": True
-                        if concept.get("dep") and concept["dep"]["value"] == "true"
-                        else False,
-                    }
-                    for concept in sparql_result[1]
-                ]
-            else:
-                return False
-
-        def render(self):
-            logging.info(f"profile is: {self.profile}")
-            current_profile = self.profiles[self.profile]
-            alt_profile_tokens = [alt["token"] for alt in self.alt_profiles.values()]
-            
-            if self.profile == "nvs":
-                if self.mediatype == "text/html":
-                    collection = self._get_collection()
-                    collection["concepts"] = self._get_concepts()
-
-                    if len(collection["concepts"]) == 0:
-                        pass
-                    elif not collection["concepts"]:
-                        return templates.TemplateResponse(
-                            "error.html",
-                            {
-                                "request": request,
-                                "title": "DB Error",
-                                "status": "500",
-                                "message": "There was an error with accessing the Triplestore",
-                            },
-                        )
-                    return templates.TemplateResponse(
-                        "collection.html",
-                        {
-                            "request": request,
-                            "uri": self.instance_uri,
-                            "collection": collection,
-                            "profile_token": self.profile,
-                            "alt_profiles": self.alt_profiles
-                        },
-                    )
-                elif self.mediatype in RDF_MEDIATYPES:
-                    # Get the term for the collection query WHERE clause to filter or accepted or deprecated.
-                    # This will be an empty string if neither condition is true.
-                    acc_dep_term = acc_dep_map.get(acc_dep_or_concept).replace("?c", "?m")
-                    query = get_collection_query(current_profile, self.instance_uri, self.ontologies, acc_dep_term)
-                    logging.info(query)
-                    return self._render_sparql_response_rdf(sparql_construct(query, self.mediatype))
-            elif self.profile == "dd":
-                q = """
-                    PREFIX dcterms: <http://purl.org/dc/terms/>
-                    PREFIX skos: <http://www.w3.org/2004/02/skos/core#>
-                    SELECT DISTINCT ?c ?pl
-                    WHERE {
-                        <xxx> skos:member ?c .
-                        acc_dep
-                        ?c skos:prefLabel ?pl .
-                    }
-                    ORDER BY ?pl                
-                    """.replace(
-                        "xxx", self.instance_uri
-                    ).replace(
-                        "acc_dep", acc_dep_map.get(acc_dep_or_concept)
-                    )
-                r = sparql_query(q)
-                return JSONResponse(
-                    [
-                        {"uri": x["c"]["value"], "prefLabel": x["pl"]["value"]}
-                        for x in r[1]
-                    ]
-                )
-            elif self.profile == "skos":
-                q = """
-                    PREFIX skos: <http://www.w3.org/2004/02/skos/core#>                    
-                    CONSTRUCT {
-                        <xxx> 
-                            a skos:Collection ;
-                            skos:prefLabel ?prefLabel ;
-                            skos:definition ?description ;
-                            skos:member ?c .
-                        ?c skos:prefLabel ?c_pl .
-                    }
-                    WHERE {
-                        <xxx> 
-                            a skos:Collection ;
-                            skos:prefLabel ?prefLabel ;
-                            <http://purl.org/dc/terms/description> ?description ;
-                            skos:member ?c .
-                            acc_dep
-                        ?c skos:prefLabel ?c_pl .
-                    }
-                    ORDER BY ?prefLabel
-                    """.replace(
-                        "xxx", self.instance_uri
-                    ).replace(
-                        "acc_dep", acc_dep_map.get(acc_dep_or_concept)
-                    )
-                return self._render_sparql_response_rdf(sparql_construct(q, self.mediatype))
-            elif self.profile == "vocpub":
-                q = """
-                    PREFIX dcterms: <http://purl.org/dc/terms/>
-                    PREFIX skos: <http://www.w3.org/2004/02/skos/core#>
-                    CONSTRUCT {
-                        <xxx>
-                            a skos:Collection ;
-                            skos:prefLabel ?prefLabel ;
-                            skos:definition ?description ;
-                            dcterms:creator ?creator ;
-                            dcterms:publisher ?publisher ;   
-                            dcterms:provenance "Made by NERC and maintained within the NERC Vocabulary Server" ;                            
-                            skos:member ?c .
-                            
-                        ?c skos:prefLabel ?c_pl .
-                    }
-                    WHERE {
-                        <xxx>
-                            a skos:Collection ;
-                            skos:prefLabel ?prefLabel ;
-                            dcterms:description ?description ;
-                            dcterms:creator ?creator ;
-                            dcterms:publisher ?publisher ;   
-                            skos:member ?c .
-                            acc_dep
-  
-                        ?c skos:prefLabel ?c_pl .
-                    }
-                    """.replace(
-                        "xxx", self.instance_uri
-                    ).replace(
-                        "acc_dep", acc_dep_map.get(acc_dep_or_concept)
-                    )
-                return self._render_sparql_response_rdf(sparql_construct(q, self.mediatype))
-            elif self.profile in alt_profile_tokens:
-                # Get the term for the collection query WHERE clause to filter or accepted or deprecated.
-                # This will be an empty string if neither condition is true.
-                acc_dep_term = acc_dep_map.get(acc_dep_or_concept).replace("?c", "?m")
-                query = get_collection_query(current_profile, self.instance_uri, self.ontologies, acc_dep_term)
-                logging.info(query)
-                return self._render_sparql_response_rdf(sparql_construct(query, self.mediatype))
-
-            alt = super().render()
-            if alt is not None:
-                return alt
-
-    return CollectionRenderer().render()
-
-
-@api.get("/collection/{collection_id}/current/{concept_id}/{vnum}/", **paths["/collection/{collection_id}/current/{concept_id}/{vnum}/"]['get'])
-@api.head("/collection/{collection_id}/current/{concept_id}/{vnum}/" , include_in_schema=False)
-def concept_with_version(request: Request, collection_id, concept_id, vnum: int):
-    return concept(request)
-
-
-@api.get("/collection/{collection_id}/current/{acc_dep_or_concept}", include_in_schema=False)
-@api.head("/collection/{collection_id}/current/{acc_dep_or_concept}" , include_in_schema=False)
-def collection_concept_noslash(request: Request, collection_id, acc_dep_or_concept):
-    return RedirectResponse(
-        url=f"/collection/{collection_id}/current/{acc_dep_or_concept}/"
-    )
-
-
-@api.get("/scheme/{scheme_id}" , include_in_schema=False)
-@api.get("/scheme/{scheme_id}/", include_in_schema=False)
-@api.head("/scheme/{scheme_id}" , include_in_schema=False)
-@api.head("/scheme/{scheme_id}/" , include_in_schema=False)
-def scheme_no_current(request: Request, scheme_id):
-    return RedirectResponse(url=f"/scheme/{scheme_id}/current/")
-
-
-@api.get("/scheme/{scheme_id}/current/", **paths["/scheme/{scheme_id}/current/"]["get"])
-@api.get("/scheme/{scheme_id}/current/{acc_dep}/", include_in_schema=False)
-@api.head("/scheme/{scheme_id}/current/" , include_in_schema=False)
-@api.head("/scheme/{scheme_id}/current/{acc_dep}/" , include_in_schema=False)
-def scheme(
-    request: Request, scheme_id, acc_dep: Literal["accepted", "deprecated", "all", None] = None
-):
-
-    if not exists_triple(request.url.path):
-      raise HTTPException(status_code=404)
-
-    class SchemeRenderer(Renderer):
-        def __init__(self):
-            self.instance_uri = f"{DATA_URI}/scheme/{scheme_id}/current/"
-
-            super().__init__(
-                request,
-                self.instance_uri,
-                {"nvs": nvs, "skos": skos, "vocpub": vocpub, "dd": dd},
-                "nvs",
-            )
-
-        def _render_sparql_response_rdf(self, sparql_response):
-            if sparql_response[0]:
-                return Response(
-                    '<?xml version="1.0" encoding="UTF-8"?>\n'.encode() + sparql_response[1]
-                    if "xml" in self.mediatype else sparql_response[1],
-                    headers={"Content-Type": self.mediatype}
-                )
-            else:
-                return PlainTextResponse(
-                    "There was an error obtaining the Concept RDF from the Triplestore",
-                    status_code=500,
-                )
-
-        def _get_scheme(self):
-            for scheme in cache_return(collections_or_conceptschemes="conceptschemes"):
-                if scheme["id"]["value"] == scheme_id:
-                    return scheme
-
-        def _get_concept_hierarchy(self):
-            def make_hierarchical_dicts(data):
-                children_parents = []
-                labels = {}
-
-                for d in data:
-                    child = d["c"]["value"]
-                    parent = (
-                        d["broader"]["value"] if d.get("broader") is not None else None
-                    )
-                    children_parents.append((child, parent))
-                    labels[child] = d["pl"]["value"].replace("<", "&lt;")
-
-                children_parents.sort(key=lambda x: x[0])
-                has_parent = set()
-                all_items = {}
-                for child, parent in children_parents:
-                    if parent not in all_items:
-                        all_items[parent] = {}
-                    if child not in all_items:
-                        all_items[child] = {}
-                    all_items[parent][child] = all_items[child]
-                    has_parent.add(child)
-
-                hierarchy = {}
-                for key, value in all_items.items():
-                    if key not in has_parent:
-                        hierarchy[key] = value
-                return hierarchy, labels
-
-            def make_nested_ul(hierarchy, labels):
-                html = ""
-                for k, v in hierarchy.items():
-                    if v:
-                        html += (
-                            f'<li><span class="caret"><a href="{k.replace(DATA_URI, "")}">{labels[k]}</a></span>'
-                            if k is not None
-                            else "None"
-                        )
-                        html += '<ul class="nested">'
-                        html += make_nested_ul(v, labels)
-                        html += "</ul>"
-                    else:
-                        html += (
-                            f'<li><a href="{k.replace(DATA_URI, "")}">{labels[k]}</a>'
-                            if k is not None
-                            else "None"
-                        )
-                    html += "</li>"
-                return html
-
-            q = """
-                PREFIX skos: <http://www.w3.org/2004/02/skos/core#>
-                SELECT DISTINCT ?c ?pl ?broader
-                WHERE {
-                  { 
-                    ?c skos:inScheme <xxx>  .
-                  }
-                  UNION
-                  { ?c skos:topConceptOf <xxx>  . }
-                  UNION
-                  { <xxx>  skos:hasTopConcept ?c . }
-                
-                  ?c skos:prefLabel ?pl .
-                  BIND (STRAFTER(STR(?c), ".uk") AS ?systemUri)
-                  
-                  acc_dep
-                  OPTIONAL { 
-                    ?c skos:broader ?broader .
-                    { ?broader skos:inScheme <xxx>  . }
-                    UNION
-                    { ?broader skos:topConceptOf <xxx>  . }
-                    UNION
-                    { <xxx>  skos:hasTopConcept ?broader . }
-                  }
-                  FILTER(lang(?pl) = "en" || lang(?pl) = "")                                    
-                }
-                ORDER BY ?pl
-                """.replace(
-                "xxx", self.instance_uri
-            ).replace(
-                "acc_dep", acc_dep_map.get(acc_dep)
-            )
-            try:
-                r = sparql_query(q)
-
-                if not r[0]:
-                    return None
-                else:
-                    hier = make_hierarchical_dicts(r[1])
-                    hier[1][None] = None
-                    return (
-                        '<ul class="concept-hierarchy">'
-                        + make_nested_ul(hier[0], hier[1])[23:-5]
-                    )
-            except RecursionError as e:
-                logging.warning(
-                    "Encountered a recursion limit error for {}".format(self.vocab_uri)
-                )
-                # make a flat list of concepts
-                q = """
-                    PREFIX skos: <http://www.w3.org/2004/02/skos/core#>
-                    SELECT DISTINCT ?c ?pl
-                    WHERE {{
-                        ?c skos:inScheme <{vocab_uri}> .              
-                        ?c skos:prefLabel ?pl .
-                        FILTER(lang(?pl) = "en" || lang(?pl) = "") 
-                    }}
-                    ORDER BY ?pl
-                    """.format(
-                    vocab_uri=self.instance_uri
-                )
-
-                concepts = [
-                    (concept["systemUri"]["value"], concept["pl"]["value"])
-                    for concept in sparql_query(q)
-                ]
-
-                concepts_html = "<br />".join(
-                    ['<a href="{}">{}</a>'.format(c[0], c[1]) for c in concepts]
-                )
-                return """<p><strong><em>This concept hierarchy cannot be displayed</em></strong><p>
-                            <p>The flat list of all this Scheme's Concepts is:</p>
-                            <p>{}</p>
-                        """.format(
-                    concepts_html
-                )
-
-        def render(self):
-            if self.profile == "nvs":
-                if self.mediatype == "text/html":
-                    scheme = self._get_scheme()
-                    scheme["concept_hierarchy"] = self._get_concept_hierarchy()
-
-                    if not scheme["concept_hierarchy"]:
-                        return templates.TemplateResponse(
-                            "error.html",
-                            {
-                                "request": request,
-                                "title": "DB Error",
-                                "status": "500",
-                                "message": "There was an error with accessing the Triplestore",
-                            },
-                        )
-
-                    return templates.TemplateResponse(
-                        "scheme.html",
-                        {
-                            "request": request,
-                            "uri": self.instance_uri,
-                            "scheme": scheme,
-                            "profile_token": "nvs",
-                        },
-                    )
-                elif self.mediatype in RDF_MEDIATYPES:
-                    q = """
-                        PREFIX dcterms: <http://purl.org/dc/terms/>
-                        PREFIX owl: <http://www.w3.org/2002/07/owl#>
-                        PREFIX skos: <http://www.w3.org/2004/02/skos/core#>
-                        PREFIX xsd: <http://www.w3.org/2001/XMLSchema#>
-                        
-                        CONSTRUCT {
-                            <xxx> ?p ?o .
-                            
-                            ?c skos:inScheme <xxx> .
-                            
-                            ?c a skos:Concept ;
-                                 skos:prefLabel ?pl ;
-                                 skos:definition ?def ;
-                                 dcterms:date ?date ;
-                                 skos:broader ?broader                                  
-                            .
-                            
-                            ?broader skos:narrower ?c .
-                        }
-                        WHERE {
-                            <xxx> ?p ?o .
-                            
-                            {
-                                ?c skos:inScheme <xxx> .
-                            }
-                            {
-                                ?c a skos:Concept ;
-                                     skos:prefLabel ?pl ;
-                                     skos:definition ?def ;
-                                     dcterms:date ?xdate ;
-                                .
-                                
-                                BIND (STRDT(REPLACE(STRBEFORE(?xdate, "."), " ", "T"), xsd:dateTime) AS ?date)
-                                
-                                FILTER(lang(?pl) = "en" || lang(?pl) = "") 
-                            }
-                                
-                            acc_dep
-                            
-                            OPTIONAL {
-                                {
-                                    ?c skos:broader ?broader .
-                                    ?broader skos:inScheme <xxx> .
-                                }
-                                UNION 
-                                {
-                                    ?broader skos:narrower ?c .
-                                    ?broader skos:inScheme <xxx> .
-                                }
-                            }                            
-                        }
-                        ORDER BY ?pl
-                        """.replace(
-                        "xxx", self.instance_uri
-                    ).replace(
-                        "acc_dep", acc_dep_map.get(acc_dep)
-                    )
-                    return self._render_sparql_response_rdf(sparql_construct(q, self.mediatype))
-            elif self.profile == "dd":
-                q = """
-                    PREFIX dcterms: <http://purl.org/dc/terms/>
-                    PREFIX skos: <http://www.w3.org/2004/02/skos/core#>
-                    SELECT DISTINCT ?c ?pl ?b
-                    WHERE {
-                        ?c skos:inScheme <xxx> ;
-                           skos:prefLabel ?pl .
-
-                        OPTIONAL {
-                            ?b skos:inScheme <xxx> .
-                            ?c skos:broader ?b .
-                        }
-                        
-                        FILTER(lang(?pl) = "en" || lang(?pl) = "")
-                    }
-                    ORDER BY ?pl                
-                    """.replace(
-                    "xxx", self.instance_uri
-                )
-                r = sparql_query(q)
-                return JSONResponse(
-                    [
-                        {
-                            "uri": x["c"]["value"],
-                            "prefLabel": x["pl"]["value"],
-                            "broader": x["b"]["value"],
-                        }
-                        if x.get("b") is not None
-                        else {"uri": x["c"]["value"], "prefLabel": x["pl"]["value"]}
-                        for x in r[1]
-                    ]
-                )
-            elif self.profile == "skos":
-                q = """
-                    PREFIX dcterms: <http://purl.org/dc/terms/>
-                    PREFIX skos: <http://www.w3.org/2004/02/skos/core#>
-                    CONSTRUCT {
-                        <xxx>
-                          a skos:ConceptScheme ;
-                          skos:prefLabel ?pl ;
-                          skos:defintion ?def ;  
-                          skos:hasTopConcept ?tc ;
-                        .
-                        ?c a skos:Concept ;
-                           skos:prefLabel ?c_pl ;
-                           skos:defintion ?c_def ; 
-                           skos:broader ?broader ;
-                           skos:inScheme <xxx> ;
-                        .
-                        
-                        ?tc skos:topConceptOf <xxx> .  
-                        ?broader skos:narrower ?c .                        
-                    }
-                    WHERE {
-                        <xxx>
-                          skos:prefLabel ?pl ;
-                          dcterms:description ?def ;  
-                          skos:hasTopConcept ?tc ;
-                        .
-                    
-                        { ?c skos:inScheme <xxx> }
-                        UNION
-                        { ?c skos:topConceptOf <xxx> }
-                        UNION
-                        { <xxx>  skos:hasTopConcept ?c }
-                    
-                        ?c 
-                            skos:prefLabel ?c_pl ;
-                            skos:definition ?c_def ;
-                        .
-                    
-                        BIND (STRAFTER(STR(?c), ".uk") AS ?systemUri)
-                        
-                        OPTIONAL { 
-                            ?c skos:broader ?broader .
-                            { ?broader skos:inScheme <xxx>  . }
-                            UNION
-                            { ?broader skos:topConceptOf <xxx>  . }
-                            UNION
-                            { <xxx>  skos:hasTopConcept ?broader . }
-                        }
-                        FILTER(lang(?pl) = "en" || lang(?pl) = "")
-                        FILTER(lang(?def) = "en" || lang(?def) = "")
-                        FILTER(lang(?c_pl) = "en" || lang(?c_pl) = "")
-                        FILTER(lang(?c_def) = "en" || lang(?c_def) = "")
-                    }
-                    ORDER BY ?pl
-                    """.replace(
-                    "xxx", self.instance_uri
-                )
-                return self._render_sparql_response_rdf(sparql_construct(q, self.mediatype))
-            elif self.profile == "vocpub":
-                q = """
-                    PREFIX dcterms: <http://purl.org/dc/terms/>
-                    PREFIX skos: <http://www.w3.org/2004/02/skos/core#>
-                    PREFIX xsd: <http://www.w3.org/2001/XMLSchema#>
-                    
-                    CONSTRUCT {
-                        <xxx>
-                          a skos:ConceptScheme ;
-                          skos:prefLabel ?pl ;
-                          skos:defintion ?def ;  
-                          skos:hasTopConcept ?tc ;
-                          dcterms:creator ?creator ;
-                          dcterms:publisher ?publisher ;
-                          dcterms:modified ?modified ;
-                          dcterms:provenance "Made by NERC and maintained within the NERC Vocabulary Server" ; 
-                        .                        
-                        
-                        ?c a skos:Concept ;
-                           skos:prefLabel ?c_pl ;
-                           skos:defintion ?c_def ; 
-                           skos:broader ?broader ;
-                           skos:inScheme <xxx> ;
-                        .
-
-                        ?tc skos:topConceptOf <xxx> .  
-                        ?broader skos:narrower ?c .                        
-                    }
-                    WHERE {
-                        <xxx>
-                          skos:prefLabel ?pl ;
-                          dcterms:description ?def ;  
-                          skos:hasTopConcept ?tc ;
-                          dcterms:publisher ?publisher ;
-                          dcterms:date ?m ;                                                    
-                        .
-                        BIND (STRDT(REPLACE(STRBEFORE(?m, "."), " ", "T"), xsd:dateTime) AS ?modified)
-                        
-                        { ?c skos:inScheme <xxx> }
-                        UNION
-                        { ?c skos:topConceptOf <xxx> }
-                        UNION
-                        { <xxx>  skos:hasTopConcept ?c }
-
-                        ?c 
-                            skos:prefLabel ?c_pl ;
-                            skos:definition ?c_def ;
-                        .
-
-                        BIND (STRAFTER(STR(?c), ".uk") AS ?systemUri)
-
-                        OPTIONAL { 
-                            ?c skos:broader ?broader .
-                            { ?broader skos:inScheme <xxx>  . }
-                            UNION
-                            { ?broader skos:topConceptOf <xxx>  . }
-                            UNION
-                            { <xxx>  skos:hasTopConcept ?broader . }
-                        }
-                        FILTER(lang(?pl) = "en" || lang(?pl) = "")
-                        FILTER(lang(?def) = "en" || lang(?def) = "")
-                        FILTER(lang(?c_pl) = "en" || lang(?c_pl) = "")
-                        FILTER(lang(?c_def) = "en" || lang(?c_def) = "")
-                    }
-                    ORDER BY ?pl
-                    """.replace(
-                    "xxx", self.instance_uri
-                )
-                return self._render_sparql_response_rdf(sparql_construct(q, self.mediatype))
-
-            alt = super().render()
-            if alt is not None:
-                return alt
-
-    return SchemeRenderer().render()
-
-
-@api.get("/scheme/{scheme_id}/current/{acc_dep}", include_in_schema=False)
-@api.head("/scheme/{scheme_id}/current/{acc_dep}" , include_in_schema=False)
-def scheme_concept_noslash(request: Request, scheme_id, acc_dep):
-    return RedirectResponse(url=f"/scheme/{scheme_id}/current/{acc_dep}/")
 
 
 @api.get("/standard_name/", include_in_schema=False)
@@ -1954,11 +1009,7 @@ def mapping(request: Request):
     return MappingRenderer().render()
 
 
-### Add the routes to the API
-api.include_router(about_page.router)
-api.include_router(contact_page.router)
-api.include_router(collection_page.router)
-api.include_router(login.router)
+
 
 
 @api.get("/.well_known/",include_in_schema=False)
